Include start/end zones in map. Include size of the map. Include the full map in the Game State
