+++ spacerace-server/src/components/ship.rs
@@ -1,4 +1,6 @@
 use bevy::prelude::Component;
+use serde::{Deserialize, Serialize};
+use uuid::Uuid;
 
 #[derive(Component)]
 pub struct ControllableShip {
@@ -6,4 +8,13 @@ pub struct ControllableShip {
 
     pub impulse: f32,
     pub torque_impulse: f32,
+}
+
+#[derive(Debug, Serialize, Deserialize, Clone)]
+pub struct Ship {
+    pub id: Uuid,
+    pub position: (f32, f32),
+    pub velocity: (f32, f32),
+    pub orientation: f32,
+    pub angular_velocity: f32,
 }
